refactor: 주문과 관련된 도메인을 별도의 패키지로 분리
