+++ src/main/java/kitchenpos/application/OrderService.java
@@ -7,12 +7,12 @@ import kitchenpos.application.exception.MenuNotFoundException;
 import kitchenpos.application.exception.OrderNotFoundException;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.repository.OrderRepository;
+import kitchenpos.domain.order.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.ui.dto.request.CreateOrderLineItemRequest;
 import kitchenpos.ui.dto.request.CreateOrderRequest;
