+++ src/main/java/kitchenpos/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION;
