+++ src/main/java/kitchenpos/domain/order/repository/OrderRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.domain.order.repository;
 
 import java.util.List;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
