+++ src/main/java/kitchenpos/ui/dto/response/CreateOrderResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.ui.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 
 public class CreateOrderResponse {
