+++ src/main/java/kitchenpos/ui/dto/response/UpdateOrderLineItemResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.dto.response;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 
 public class UpdateOrderLineItemResponse {
