+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -12,16 +12,16 @@ import kitchenpos.config.IntegrationTest;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menu.MenuProduct;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.exception.InvalidNumberOfGuestsException;
 import kitchenpos.domain.exception.InvalidOrderStatusCompletionException;
 import kitchenpos.domain.menugroup.repository.MenuGroupRepository;
 import kitchenpos.domain.menu.repository.MenuRepository;
-import kitchenpos.repository.OrderRepository;
+import kitchenpos.domain.order.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.domain.product.repository.ProductRepository;
 import kitchenpos.repository.TableGroupRepository;
